+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.2.0] - 2024-11-20
+
+- Removed custom error type `ChibiHashError`
+- Added support for streaming hashing through `StreamingChibiHasher`
+
 ## [v0.1.2] - 2024-11-19
 
 - Bit rotation the Rust way
@@ -25,7 +30,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added benchmarks
 - Released a crate
 
-[Unreleased]: https://github.com/thevilledev/ChibiHash-rs/compare/v0.1.2...HEAD
+[Unreleased]: https://github.com/thevilledev/ChibiHash-rs/compare/v0.2.0...HEAD
+[v0.2.0]: https://github.com/thevilledev/ChibiHash-rs/compare/v0.1.2...v0.2.0
 [v0.1.2]: https://github.com/thevilledev/ChibiHash-rs/compare/v0.1.1...v0.1.2
 [v0.1.1]: https://github.com/thevilledev/ChibiHash-rs/compare/v0.1.0...v0.1.1
 [v0.1.0]: https://github.com/thevilledev/ChibiHash-rs/releases/tag/v0.1.0